Revised Index.js in routes and outside of server folder
