+++ zoosta/server/routes/index.js
@@ -1,23 +1,8 @@
-const express = require('express');
-const db = require('./config/connection');
-const routes = require('./routes');
+const router = require('express').Router();
+const apiRoutes = require('./api');
 
-const cwd = process.cwd();
+router.use('/api', apiRoutes);
 
-const PORT = process.env.PORT || 3001;
-const app = express();
+router.use((req, res) => res.send('Wrong route!'));
 
-// Note: not necessary for the Express server to function. This just helps indicate what activity's server is running in the terminal.
-const activity = cwd.includes('01-Activities')
-  ? cwd.split('/01-Activities/')[1]
-  : cwd;
-
-app.use(express.urlencoded({ extended: true }));
-app.use(express.json());
-app.use(routes);
-
-db.once('open', () => {
-  app.listen(PORT, () => {
-    console.log(`API server for ${activity} running on port ${PORT}!`);
-  });
-});
+module.exports = router;
